Problem 3: 12% performance gain thanks to uints.

+++ problem3/problem3.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 )
 
-func compute(input int) int {
-	for i := 2; i < input; i++ {
+func compute(input uint) uint {
+	for i := uint(2); i < input; i++ {
 		if input%i == 0 {
 			input /= i
 		}
